Add a replace entry for `phpseclib/phpseclib`
The `phpseclib/phpseclib` dependency of `google/apiclient` is not PHP 7+ compatible and not used by Site Kit, so it should be removed. Since it is an inherited dependency the removal needs to be done via Composer.

The `composer.json` can have `replace` entries to specify dependencies fulfilled by the current project, which will not be installed by Composer. The intention is to use forked or self-maintained versions of a dependency but, in this case, can be used to skip the dependency entirely. See [the `composer.json` replace doc](https://getcomposer.org/doc/04-schema.md#replace) for details.