Update and simplify activation failure tests.
These tests were simplified by receiving modules directly into the store rather than mocking the fetch request triggered in the resolver of getModules and awaiting its resolution in the test since these tests aren't concerned with these resolvers.

The authentication request is only made if the activation request is successful now so the fetch mock for this was removed as well and the expected number of fetches reduced to 1 for both.