Update `propTypes` for the `CTA` component
Register the missing property `ctaLinkExternal` in order to allow
callers to identify available properties, it sets a default value of
`false` as well in order to be backwards compatible as if the value
is not set `false` would be used instead for existing components where
`CTA` component is used.

This work was based out of the ticket #4286.