Add filter to toggle modules programmatically.
A new filter was added `googlesitekit_available_modules`
the filter allow customizing programmatically which
modules are present on the site.

Each module now has a unique key in order to
identify each module during the filter just by
passing a slug of the module.

This PR follows the IB from ticket #3993.