Add optional nonInteractive prop to Menu.
Adds the mdc-list--non-interactive class to the menu.

Note: This class has been removed in newer versions of @material/list,
but this should be OK as we've moving to @mui for the rebrand and should
not need to update @material/list.

See https://github.com/material-components/material-components-web/tree/v2.3.0/packages/mdc-list#css-classes .