Fix a TourTooltips test but without addressing the underlying issue.
A proper fix would be to update the tests so the Floater is visible.

Presently the floater div is visibility: hidden as seen in the debug
output:

        <div
          class="__floater"
          style="display: inline-block; filter: drop-shadow(rgba(60, 64, 67, 0.3) 0px 1px 2px) drop-shadow(rgba(60, 64, 67, 0.15) 0px 2px 6px); max-width: 100%; opacity: 0; position: absolute; transition: opacity 0.3s; visibility: hidden; z-index: 20000;"
          x-placement="top"
        >

This has not changed but it looks like the bump to JSDom 16 means the visibility is now being propagated correctly causing an issue for getByRole (even when using hidden: true for getByRole it's still a problem for the accessible name lookup which also checks visibility).