Avoid rendering same SVG multiple times to prevent duplicate ID issue.
When rendering SVGs which use IDs to refer to defs, problems can occur
when multiple instances of the SVG are rendered in the page, as this
results in duplicate IDs and one instance of an SVG can have the ID its
looking up resolved to another instance, causing display problems.

That is what the Babel plugin babel-plugin-react-inline-svg-unique-id is
designed to solve, but unfortunately it requires React 17. For now,
avoid the problem by not rendering the same SVG multiple times in a
page.