Apply fix to avoid "changing key" hack.
Turns out setting the `checked` property on the `md-checkbox` doesn't
currently reflect properly to the `checked` property on the inner
`input`. As a fix, we set `input.checked` manually.

This improves rendering performance, ensures the checkbox animation is
not cut off, and prevents the checkbox from losing its focus when
toggled.