Fix survey trigger.
The value of `usingProxy` was undefined in `useMount` due to the value
coming from a resolver which is now async due to the @wordpress/data
update. This meant `triggerSurvey` was never being executed.

Switched to a `useEffect` which will only fire `triggerSurvey` once, when
`usingProxy` becomes true (or the `triggerID` / `ttl` props change, but
this is not expected in current usage).