Remove test for unsupported behaviour.
Having updated `eslint-plugin-jsdoc`, the `jsdoc-requires-since` rule no
longer warns for whitespace before the `@since` version.

This is because leading whitespace is now trimmed from
`tag.description`.

It would be possible to restore this functionality by parsing
`sinceTags[ 0 ].source`, if so desired. However it doesn't seem
essential to the rule set as currently defined in
`SINCE_VALIDATION_RULES`.